Refactoring the stomp client and tests a bit
The main goal of this refactor is to tell the client to
be resilient to failures (configured at initialization time),
and to send all messages as persistent messages (configured
for each message).

In the process, the client now parses the queue source URI
and handles each argument separately.  The tests are more
thorough, also.

Signed-off-by: Luke Kanies <luke@madstop.com>